# openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/employees/{id}":{"get":{"tags":["employee-controller"],"summary":"Get detailed employee information by ID","description":"Retrieves detailed information about an employee, including their personal information,\nsupervisor, last update timestamp, and the number of subordinates (direct and indirect).\nThe employee retrieval and subordinate counting are executed in parallel for optimized performance.","operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EmployeeResponse"}}}}}},"put":{"tags":["employee-controller"],"summary":"Update an employee","description":"Updates an existing employee identified by its ID.\n The payload must include valid employee attributes.\nThe response returns the updated employee with its last update timestamp.","operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployeeRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EmployeeResponse"}}}}}},"delete":{"tags":["employee-controller"],"summary":"Delete an employee","description":"Deletes an existing employee identified by its ID. \nIf the employee does not exist, a 404 Not Found is returned.","operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/employees/password":{"put":{"tags":["employee-controller"],"summary":"Update the authenticated user's password","description":"Allows the currently authenticated user to update their password.\nThe username is extracted from the JWT token; no username is required in the request body.","operationId":"updatePassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PasswordUpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/login":{"post":{"tags":["auth-controller"],"summary":"User login","description":"Returns a JWT valid for 5 minutes","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}}}}},"/employees":{"get":{"tags":["employee-controller"],"summary":"Get all employees","description":"Returns a list of all employees in the system, including their basic details and the\ntimestamp of the last update. This endpoint does not calculate subordinates.","operationId":"getAll","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/EmployeeResponse"}}}}}}},"post":{"tags":["employee-controller"],"summary":"Create or update an employee","description":"Creates a new employee or updates an existing one based on the provided JSON payload.\nIf an ID is included, the existing record is updated; otherwise, a new employee is created.\nThe response includes the persisted employee data and the timestamp of the last update.","operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployeeRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EmployeeResponse"}}}}}}},"/logout":{"delete":{"tags":["auth-controller"],"summary":"User logout","description":"Invalidates current JWT","operationId":"logout","responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"EmployeeRequest":{"required":["email","name","userName"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"userName":{"type":"string"},"name":{"type":"string"},"email":{"type":"string"},"supervisorId":{"type":"integer","format":"int64"},"password":{"type":"string"}}},"EmployeeResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"email":{"type":"string"},"supervisorId":{"type":"integer","format":"int64"},"updatedAt":{"type":"string","format":"date-time"},"subordinatesCount":{"type":"integer","format":"int32"}}},"PasswordUpdateRequest":{"required":["newPassword","oldPassword"],"type":"object","properties":{"oldPassword":{"type":"string"},"newPassword":{"type":"string"}}},"LoginRequest":{"required":["password","username"],"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}},"LoginResponse":{"type":"object","properties":{"token":{"type":"string"},"expiresIn":{"type":"integer","format":"int64"}}}}}}